+++ src/app/workers/tsconfig.json
@@ -1,0 +1,15 @@
+{
+	"extends": "../../../tsconfig.json",
+	"compilerOptions": {
+		"esModuleInterop": true,
+		"outDir": "./public/js/workers",
+		"lib": [
+			"WebWorker",
+			"Webworker.Iterable",
+			"WebWorker.AsyncIterable",
+			"WebWorker.ImportScripts"
+		],
+		"noEmit": false
+	},
+	"include": ["**/*.ts"]
+}
